Refactor: Refactored map indexes
