+++ src/frontend/pages/watchlater/WatchLaterPage.jsx
@@ -1,19 +1,17 @@
 import React, { useEffect } from "react";
+import "./watchlaterpage.css";
 import { Link, useNavigate } from "react-router-dom";
-import { Sidebar } from "../../components/componentExport";
-import { HorizontalVideoCard } from "../../components/HorizontalVideoCard/HorizontalVideoCard";
-import { useAuth } from "../../contexts/contextExport";
-import { useDataStore } from "../../contexts/contextExport";
+import { Sidebar, HorizontalVideoCard } from "../../components/componentExport";
+import { useAuth, useDataStore } from "../../contexts/contextExport";
 import { deleteWatchLaterHandler } from "../../helperfunctions/watchLaterHandler";
 import Watchlater from "../../images/Watchlater.svg";
-import "./watchlaterpage.css";
 
 export const WatchLaterPage = () => {
   const { isLoggedIn } = useAuth();
-
   const navigate = useNavigate();
-  const { dataStoreState } = useDataStore();
-  const { videos } = dataStoreState;
+  const {
+    dataStoreState: { videos },
+  } = useDataStore();
   const watchLaterVideos = videos?.filter((item) => item.isInWatchLater);
 
   useEffect(() => {
@@ -39,9 +37,10 @@ export const WatchLaterPage = () => {
           </div>
           <div className="fv-videos">
             {watchLaterVideos.length >= 1 ? (
-              watchLaterVideos.map((item) => {
+              watchLaterVideos.map((item, index) => {
                 return (
                   <HorizontalVideoCard
+                    key={index}
                     deleteHandler={deleteWatchLaterHandler}
                     videoItem={item}
                   />
